+++ frontend/package.json
@@ -44,7 +44,6 @@
     "@types/node": "^20.12.6",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
-    "@types/react-table": "^7.7.20",
     "@vitejs/plugin-react": "^4.2.1",
     "@vitest/coverage-v8": "^1.4.0",
     "@vitest/ui": "^1.2.2",
@@ -62,7 +61,6 @@
     "prettier": "^3.2.5",
     "prettier-plugin-organize-imports": "^3.2.4",
     "pretty-quick": "^4.0.0",
-    "react-table": "^7.8.0",
     "recharts": "^2.12.6",
     "sass": "^1.74.1",
     "typescript": "^5.4.4",
