+++ frontend/src/components/forms/MovieUploadForm.tsx
@@ -20,7 +20,7 @@ import { ColumnDef } from "@tanstack/react-table";
 import { isString } from "lodash";
 import { useMovieSubtitleModification } from "@/apis/hooks";
 import { Action, Selector } from "@/components/inputs";
-import NewSimpleTable from "@/components/tables/NewSimpleTable";
+import SimpleTable from "@/components/tables/SimpleTable";
 import TextPopover from "@/components/TextPopover";
 import { useModals, withModal } from "@/modules/modals";
 import { task, TaskGroup } from "@/modules/task";
@@ -291,10 +291,7 @@ const MovieUploadForm: FunctionComponent<Props> = ({
       })}
     >
       <Stack className="table-long-break">
-        <NewSimpleTable
-          columns={columns}
-          data={form.values.files}
-        ></NewSimpleTable>
+        <SimpleTable columns={columns} data={form.values.files}></SimpleTable>
         <Divider></Divider>
         <Button type="submit">Upload</Button>
       </Stack>
