+++ frontend/src/components/tables/BaseTable.tsx
@@ -1,11 +1,17 @@
-import { ReactNode, useMemo } from "react";
-import { HeaderGroup, Row, TableInstance } from "react-table";
+import React, { ReactNode, useMemo } from "react";
 import { Box, Skeleton, Table, Text } from "@mantine/core";
+import {
+  flexRender,
+  Header,
+  Row,
+  Table as TableInstance,
+} from "@tanstack/react-table";
 import { useIsLoading } from "@/contexts";
 import { usePageSize } from "@/utilities/storage";
-import styles from "./BaseTable.module.scss";
+import styles from "@/components/tables/BaseTable.module.scss";
 
-export type BaseTableProps<T extends object> = TableInstance<T> & {
+export type BaseTableProps<T extends object> = {
+  instance: TableInstance<T>;
   tableStyles?: TableStyleProps<T>;
 };
 
@@ -15,60 +21,57 @@ export interface TableStyleProps<T extends object> {
   placeholder?: number;
   hideHeader?: boolean;
   fixHeader?: boolean;
-  headersRenderer?: (headers: HeaderGroup<T>[]) => JSX.Element[];
-  rowRenderer?: (row: Row<T>) => Nullable<JSX.Element>;
+  headersRenderer?: (headers: Header<T, unknown>[]) => React.JSX.Element[];
+  rowRenderer?: (row: Row<T>) => Nullable<React.JSX.Element>;
 }
 
 function DefaultHeaderRenderer<T extends object>(
-  headers: HeaderGroup<T>[],
-): JSX.Element[] {
-  return headers.map((col) => (
-    <Table.Th style={{ whiteSpace: "nowrap" }} {...col.getHeaderProps()}>
-      {col.render("Header")}
+  headers: Header<T, unknown>[],
+): React.JSX.Element[] {
+  return headers.map((header) => (
+    <Table.Th style={{ whiteSpace: "nowrap" }} key={header.id}>
+      {flexRender(header.column.columnDef.header, header.getContext())}
     </Table.Th>
   ));
 }
 
-function DefaultRowRenderer<T extends object>(row: Row<T>): JSX.Element | null {
+function DefaultRowRenderer<T extends object>(
+  row: Row<T>,
+): React.JSX.Element | null {
   return (
-    <Table.Tr {...row.getRowProps()}>
-      {row.cells.map((cell) => (
-        <Table.Td {...cell.getCellProps()}>{cell.render("Cell")}</Table.Td>
+    <Table.Tr key={row.id}>
+      {row.getVisibleCells().map((cell) => (
+        <Table.Td key={cell.id}>
+          {flexRender(cell.column.columnDef.cell, cell.getContext())}
+        </Table.Td>
       ))}
     </Table.Tr>
   );
 }
 
 export default function BaseTable<T extends object>(props: BaseTableProps<T>) {
-  const {
-    headerGroups,
-    rows: tableRows,
-    page: tablePages,
-    prepareRow,
-    getTableProps,
-    getTableBodyProps,
-    tableStyles,
-  } = props;
+  const { instance, tableStyles } = props;
 
   const headersRenderer = tableStyles?.headersRenderer ?? DefaultHeaderRenderer;
   const rowRenderer = tableStyles?.rowRenderer ?? DefaultRowRenderer;
 
   const colCount = useMemo(() => {
-    return headerGroups.reduce(
-      (prev, curr) => (curr.headers.length > prev ? curr.headers.length : prev),
-      0,
-    );
-  }, [headerGroups]);
+    return instance
+      .getHeaderGroups()
+      .reduce(
+        (prev, curr) =>
+          curr.headers.length > prev ? curr.headers.length : prev,
+        0,
+      );
+  }, [instance]);
 
-  // Switch to usePagination plugin if enabled
-  const rows = tablePages ?? tableRows;
-
-  const empty = rows.length === 0;
+  const empty = instance.getRowCount() === 0;
 
   const pageSize = usePageSize();
   const isLoading = useIsLoading();
 
   let body: ReactNode;
+
   if (isLoading) {
     body = Array(tableStyles?.placeholder ?? pageSize)
       .fill(0)
@@ -88,27 +91,22 @@ export default function BaseTable<T extends object>(props: BaseTableProps<T>) {
       </Table.Tr>
     );
   } else {
-    body = rows.map((row) => {
-      prepareRow(row);
+    body = instance.getRowModel().rows.map((row) => {
       return rowRenderer(row);
     });
   }
 
   return (
     <Box className={styles.container}>
-      <Table
-        className={styles.table}
-        striped={tableStyles?.striped ?? true}
-        {...getTableProps()}
-      >
+      <Table className={styles.table} striped={tableStyles?.striped ?? true}>
         <Table.Thead hidden={tableStyles?.hideHeader}>
-          {headerGroups.map((headerGroup) => (
-            <Table.Tr {...headerGroup.getHeaderGroupProps()}>
+          {instance.getHeaderGroups().map((headerGroup) => (
+            <Table.Tr key={headerGroup.id}>
               {headersRenderer(headerGroup.headers)}
             </Table.Tr>
           ))}
         </Table.Thead>
-        <Table.Tbody {...getTableBodyProps()}>{body}</Table.Tbody>
+        <Table.Tbody>{body}</Table.Tbody>
       </Table>
     </Box>
   );
