+++ frontend/src/components/tables/SimpleTable.tsx
@@ -1,23 +1,61 @@
-import { PluginHook, TableInstance, TableOptions, useTable } from "react-table";
-import BaseTable, { TableStyleProps } from "./BaseTable";
-import { useDefaultSettings } from "./plugins";
+import { MutableRefObject, useEffect, useMemo } from "react";
+import {
+  getCoreRowModel,
+  Row,
+  Table,
+  TableOptions,
+  useReactTable,
+} from "@tanstack/react-table";
+import BaseTable, { TableStyleProps } from "@/components/tables/BaseTable";
+import { usePageSize } from "@/utilities/storage";
 
-export type SimpleTableProps<T extends object> = TableOptions<T> & {
-  plugins?: PluginHook<T>[];
-  instanceRef?: React.MutableRefObject<TableInstance<T> | null>;
+export type SimpleTableProps<T extends object> = Omit<
+  TableOptions<T>,
+  "getCoreRowModel"
+> & {
+  instanceRef?: MutableRefObject<Table<T> | null>;
   tableStyles?: TableStyleProps<T>;
+  onRowSelectionChanged?: (selectedRows: Row<T>[]) => void;
+  onAllRowsExpandedChanged?: (isAllRowsExpanded: boolean) => void;
 };
 
 export default function SimpleTable<T extends object>(
   props: SimpleTableProps<T>,
 ) {
-  const { plugins, instanceRef, tableStyles, ...options } = props;
+  const {
+    instanceRef,
+    tableStyles,
+    onRowSelectionChanged,
+    onAllRowsExpandedChanged,
+    ...options
+  } = props;
 
-  const instance = useTable(options, useDefaultSettings, ...(plugins ?? []));
+  const pageSize = usePageSize();
+
+  const instance = useReactTable({
+    ...options,
+    getCoreRowModel: getCoreRowModel(),
+    autoResetPageIndex: false,
+    autoResetExpanded: false,
+    pageCount: pageSize,
+  });
 
   if (instanceRef) {
     instanceRef.current = instance;
   }
 
-  return <BaseTable tableStyles={tableStyles} {...instance}></BaseTable>;
+  const selectedRows = instance.getSelectedRowModel().rows;
+  const memoizedRows = useMemo(() => selectedRows, [selectedRows]);
+
+  const isAllRowsExpanded = instance.getIsAllRowsExpanded();
+
+  useEffect(() => {
+    onRowSelectionChanged?.(memoizedRows);
+  }, [onRowSelectionChanged, memoizedRows]);
+
+  useEffect(() => {
+    onAllRowsExpandedChanged?.(isAllRowsExpanded);
+  }, [onAllRowsExpandedChanged, isAllRowsExpanded]);
+
+  return <BaseTable tableStyles={tableStyles} instance={instance}></BaseTable>;
 }
