--- frontend/src/components/tables/plugins/index.ts
@@ -1,2 +1,0 @@
-export { default as useCustomSelection } from "./useCustomSelection";
-export { default as useDefaultSettings } from "./useDefaultSettings";
